requirements.txt
python-telegram-bot
python-telegram-bot[webhooks]
langchain-anthropic
langchain-community
langchain
python-dotenv
speechrecognition
pydub
flask